enable updateTargetServer command
and update README accordingly
also change case of the name of various command files.
