+++ lib/cli.js
@@ -1,6 +1,4 @@
 #!/usr/bin/env node
-/* jshint node: true  */
-'use strict';
 
 var main = require('./main');
 var options = require('./options');
@@ -64,13 +62,13 @@ function runCommand() {
       }
       process.exit(6);
     }
-    
+
     if (!opts.json && commandModule.format) {
       console.log(commandModule.format(result));
     } else {
       console.log(JSON.stringify(result));
     }
-    
+
     process.exit(0);
   });
 }
