+++ lib/commands/commands.js
@@ -30,6 +30,10 @@ const Commands = {
     description: 'List Target Servers',
     load: () => require('./listTargetServers')
   },
+  updateTargetServer: {
+    description: 'Update a Target Server',
+    load: () => require('./updateTargetServer')
+  },
   listProxies: {
     description: 'List Proxies',
     load: () => require('./listProxies')
@@ -76,7 +80,7 @@ const Commands = {
   },
   listdeployments: {
     description: 'List Deployments',
-    load: () => require('./listdeployments')
+    load: () => require('./listDeployments')
   },
   undeploy: {
     description: 'Undeploy Proxy or Node.js Application',
