+++ lib/commands/deployHostedTarget.js
@@ -1,43 +1,34 @@
-/* jshint node: true  */
-'use strict';
-
 const util = require('util'),
       path = require('path'),
       async = require('async'),
-      fs = require('fs'),
+      fs = require('fs-extra'),
       mustache = require('mustache'),
       tmp = require('tmp');
 
 tmp.setGracefulCleanup();
 
-var defaults = require('../defaults');
-var options = require('../options');
-var ziputils = require('../ziputils');
-var parseDeployments = require('./parsedeployments');
-var fetchproxy = require('./fetchproxy');
-var deployproxy = require('./deployproxy');
-var createApiProxy = require('../deploycommon').createApiProxy;
-var createProxy = require('../deploycommon').createProxy;
-var deployProxy = require('../deploycommon').deployProxy;
-var unzipProxy = require('../deploycommon').unzipProxy;
-var copyFile = require('../deploycommon').copyFile;
-var handleUploadResult = require('../deploycommon').handleUploadResult;
-var usePackedSource = require('../deploycommon').usePackedSource;
-var uploadSource = require('../deploycommon').uploadSource;
+const defaults = require('../defaults'),
+      options = require('../options'),
+      ziputils = require('../ziputils'),
+      parseDeployments = require('./parsedeployments'),
+      fetchProxy = require('./fetchProxy'),
+      deployProxy1 = require('./deployProxy'),
+      createApiProxy = require('../deploycommon').createApiProxy,
+      createProxy = require('../deploycommon').createProxy,
+      deployProxy2 = require('../deploycommon').deployProxy,
+      unzipProxy = require('../deploycommon').unzipProxy,
+      copyFile = require('../deploycommon').copyFile,
+      handleUploadResult = require('../deploycommon').handleUploadResult,
+      usePackedSource = require('../deploycommon').usePackedSource,
+      uploadSource = require('../deploycommon').uploadSource;
 
-var cleanResults = require('../utils').cleanResults;
-
-var DeploymentDelay = 60;
-var ProxyBase = 'apiproxy';
-
-// By default, do not run NPM remotely
-var DefaultResolveModules = false;
+const APP_YAML = 'app.yaml';
 
 var descriptor = defaults.defaultDescriptor({
   api: {
     name: 'API Name',
     shortOption: 'n',
-    required: false
+    required: true
   },
   environments: {
     name: 'Environments',
@@ -47,11 +38,6 @@ var descriptor = defaults.defaultDescriptor({
   directory: {
     name: 'Directory',
     shortOption: 'd',
-    required: false
-  },
-  main: {
-    name: 'Main Script',
-    shortOption: 'm',
     required: false
   },
   virtualhosts: {
@@ -67,9 +53,9 @@ var descriptor = defaults.defaultDescriptor({
     shortOption: 'i',
     toggle: true
   },
-  'resolve-modules': {
-    name: 'Resolve Modules',
-    shortOption: 'R',
+  'preserve-policies': {
+    name: 'Preserve policies from previous revision',
+    shortOption: 'P',
     toggle: true
   },
   'bundled-dependencies' : {
@@ -80,20 +66,6 @@ var descriptor = defaults.defaultDescriptor({
     name: 'Upload Modules',
     shortOption: 'U',
     toggle: true
-  },
-  'preserve-policies': {
-    name: 'Preserve policies from previous revision',
-    shortOption: 'P',
-    toggle: true
-  },
-  'wait-after-import': {
-    name: 'Wait N seconds after importing proxy before deploying',
-    shortOption: 'W',
-    required: false
-  },
-  'production': {
-    name: 'Run resolve modules with npm install --production, default true',
-    required: false
   }
 });
 module.exports.descriptor = descriptor;
@@ -106,42 +78,21 @@ module.exports.format = function(r) {
   return result;
 };
 
-
 module.exports.run = function(opts, cb) {
-  // Setup Delay for after import if set
-  opts.waitAfterImportDelay = 0;
-  if (opts['wait-after-import'] !== undefined) {
-    opts.waitAfterImportDelay = parseInt(opts['wait-after-import']);
-    if (isNaN(opts.waitAfterImportDelay)) {
-      console.error('invalid int for wait-after-import');
-      process.exit(1);
-    }
-  }
-
   if (!opts.directory) {
     opts.directory = process.cwd();
   }
-  if (!opts.main || !opts.api) {
-    try {
-      var packageFile = path.resolve(opts.directory, 'package.json');
-      var packageObj = JSON.parse(fs.readFileSync(packageFile, 'utf8'));
-      if (!opts.main) {
-        opts.main = packageObj.main;
-      }
-      if (!opts.api) {
-        opts.api = packageObj.name;
-      }
-    } catch (err) {
-      if (opts.debug) {
-        console.error('unable to read package.json', err);
-      }
-    }
-  }
-  descriptor.main.required = true;
+
   descriptor.api.required = true;
   options.validateSync(opts, descriptor);
   if (opts.debug) {
-    console.log('deploynodeapp: %j', opts);
+    console.log('deployhostedtarget: %j', opts);
+  }
+
+  try {
+    fs.statSync(path.join(opts.directory, APP_YAML))
+  } catch(err) {
+    return cb(new Error("missing required 'app.yaml'"));
   }
 
   var request = defaults.defaultRequest(opts);
@@ -152,6 +103,11 @@ module.exports.run = function(opts, cb) {
     // if preserve-policies, we do something entirely different...
     if (opts['preserve-policies'] && opts.deploymentVersion > 1) {
       return preservePoliciesRun(opts, cb);
+    }
+
+    opts.remoteNpm = true;
+    if (opts['upload-modules'] && (opts['upload-modules'] === true)) {
+      opts.remoteNpm = false;
     }
 
     var steps = [
@@ -174,7 +130,7 @@ module.exports.run = function(opts, cb) {
 
     steps = steps.concat([
       function(done) {
-        uploadSource(opts.directory, 'node', opts, request, done);
+        uploadSource(opts.directory, 'hosted', opts, request, done);
       },
       function(done) {
         createTarget(opts, request, done);
@@ -183,10 +139,7 @@ module.exports.run = function(opts, cb) {
         createProxy(opts, request, done);
       },
       function(done) {
-        runNpm(opts, request, done);
-      },
-      function(done) {
-        deployProxy(opts, request, done);
+        deployProxy2(opts, request, done);
       }
     ])
 
@@ -194,14 +147,7 @@ module.exports.run = function(opts, cb) {
     async.series(steps,
       function(err, results) {
         if (err) { return cb(err); }
-
-        results = cleanResults(results);
-
         if (opts.debug) { console.log('results: %j', results); }
-
-        if (opts['import-only']) {
-          return cb(undefined, results)
-        }
 
         async.map(Object.values(results[results.length - 1]),
           function(result, cb) {
@@ -236,24 +182,24 @@ function preservePoliciesRun(opts, cb) {
     opts.file = fetchedProxyZip;
 
     if (opts.verbose) { console.log('Downloading proxy %s revision %d', opts.name, opts.revision); }
-    fetchproxy.run(opts, function(err) {
+    fetchProxy.run(opts, function(err) {
       if (err) { return cb(err); }
 
       // set up temporary project dir
       tmp.dir({ unsafeCleanup: false }, function(err, tmpDir) {
         if (err) { return cb(err); }
 
-        unzipProxy(opts, tmpDir, 'apiproxy/resources/node', function(err) {
+        unzipProxy(opts, tmpDir, 'apiproxy/resources/hosted', function(err) {
           if (err) { return cb(err); }
 
-          // copy node files to tmpDir node directory
-          var nodeResourceDir = path.resolve(tmpDir, 'apiproxy/resources/node');
-          copyNodeSource(opts, nodeResourceDir, function(err) {
+          // copy hosted files to tmpDir hosted directory
+          var sourcesDir = path.resolve(tmpDir, 'apiproxy/resources/hosted');
+          copySources(opts, sourcesDir, function(err) {
             if (err) { return cb(err); }
 
             // deploy proxy at tmpDir
             opts.directory = tmpDir;
-            deployproxy.run(opts, cb);
+            deployProxy1.run(opts, cb);
           });
         });
       });
@@ -261,13 +207,12 @@ function preservePoliciesRun(opts, cb) {
   });
 }
 
-function copyNodeSource(opts, targetDir, cb) {
+function copySources(opts, targetDir, cb) {
 
-  if (opts.verbose) { console.log('Copying node source into proxy'); }
+  if (opts.verbose) { console.log('Copying sources into proxy'); }
 
-  // Get a list of entries, broken down by which are directories,
-  // and with special handling for the node_modules directory.
-  ziputils.enumerateDirectory(opts.directory, 'node', opts.remoteNpm, function(err, entries) {
+  // Get a list of entries, broken down by which are directories
+  ziputils.enumerateDirectory(opts.directory, 'hosted', opts.remoteNpm, function(err, entries) {
     if (err) { return cb(err); }
 
     if (opts.debug) { console.log('Directories to copy: %j', entries); }
@@ -314,32 +259,6 @@ function getDeploymentInfo(opts, request, done) {
     return;
   }
 
-  if (!fs.existsSync(path.join(opts.directory, opts.main))) {
-    // Main script might be an absolute path, so fix it up
-    opts.main = path.relative(opts.directory, opts.main);
-  }
-  if (!fs.existsSync(path.join(opts.directory, opts.main))) {
-    done(new Error(util.format('Main script file %s does not seem to exist', opts.main)));
-    return;
-  }
-  if (path.dirname(opts.main) !== '.') {
-    done(new Error(util.format('Main script file %s must be in the top level directory',
-      opts.main)));
-      return;
-  }
-
-  // Check out some specific parameters that aren't caught by the generic stuff
-  opts.remoteNpm = DefaultResolveModules;
-  if (opts['upload-modules'] && (opts['upload-modules'] === true)) {
-    opts.remoteNpm = false;
-  }
-  if (opts['resolve-modules'] && (opts['resolve-modules'] === true)) {
-    opts.remoteNpm = true;
-  }
-  if (opts.debug) {
-    console.log('Resolve NPM modules = %s', opts.remoteNpm);
-  }
-
   // Find out which revision we should be creating
   request.get(util.format('%s/v1/o/%s/apis/%s',
                opts.baseuri, opts.organization, opts.api),
@@ -372,15 +291,13 @@ function getDeploymentInfo(opts, request, done) {
   });
 }
 
-// Create a target endpoint that references the Node.js script
+// Create a target endpoint that references the Hosted Target target
 function createTarget(opts, request, done) {
   var targetDoc = mustache.render(
     '<TargetEndpoint name="default">' +
     '<PreFlow name="PreFlow"/>' +
     '<PostFlow name="PostFlow"/>' +
-    '<ScriptTarget>' +
-    '<ResourceURL>node://{{main}}</ResourceURL>' +
-    '</ScriptTarget>' +
+    '<HostedTarget/>' +
     '</TargetEndpoint>', opts);
 
   var uri = util.format('%s/v1/o/%s/apis/%s/revisions/%d/targets?name=default',
@@ -399,50 +316,4 @@ function createTarget(opts, request, done) {
   }, function(err, req, body) {
     handleUploadResult(err, req, 'targets/default.xml', done);
   });
-}
-
-function runNpm(opts, request, done) {
-  if (!opts.remoteNpm && !opts['bundled-dependencies']) {
-    done();
-  } else {
-    if (opts.verbose) {
-      console.log('Running "npm install" at Apigee. This may take several minutes.');
-    }
-
-    var installURI = util.format('%s/v1/o/%s/apis/%s/revisions/%d/npm?command=install',
-    opts.baseuri, opts.organization, opts.api, opts.deploymentVersion)
-
-    if (opts.production === 'false') {
-      installURI = util.format("%s&production=false", installURI)
-    }
-
-    if (opts.debug) {
-      installURI = util.format("%s&verbose=true", installURI)
-    }
-
-    request({
-      uri: installURI,
-      method: 'POST',
-      headers: {
-        'Accept': 'text/plain'
-      },
-      json: false
-    }, function(err, req, body) {
-      if (err) {
-        done(err);
-      } else if (req.statusCode === 200) {
-        if (opts.verbose) {
-          console.log('NPM complete.');
-          console.log(body);
-        }
-        done();
-      } else {
-        if (opts.verbose) {
-          console.log('NPM failed with %d', req.statusCode);
-          console.log(body);
-        }
-        done(new Error(util.format('NPM install failed with status code %d', req.statusCode)));
-      }
-    });
-  }
 }
