+++ lib/commands/getTargetServer.js
@@ -1,9 +1,5 @@
-/* jshint node: true  */
-'use strict';
-
 const util = require('util'),
       defaults = require('../defaults'),
-      options = require('../options'),
       command_utils = require('./command-utils');
 
 var descriptor = defaults.defaultDescriptor({
@@ -15,6 +11,7 @@ var descriptor = defaults.defaultDescriptor({
   },
   targetServerName: {
     name: 'Target Server Name',
+    shortOption: 'n',
     required: true,
     prompt: true
   }
