+++ lib/commands/listDeployments.js
@@ -1,6 +1,3 @@
-/* jshint node: true  */
-'use strict';
-
 const util = require('util'),
       defaults = require('../defaults'),
       options = require('../options'),
