+++ lib/commands/listTargetServers.js
@@ -1,9 +1,5 @@
-/* jshint node: true  */
-'use strict';
-
 const util = require('util'),
       defaults = require('../defaults'),
-      options = require('../options'),
       command_utils = require('./command-utils');
 
 var descriptor = defaults.defaultDescriptor({
