+++ tools/ci/linux/presubmit.sh
@@ -1,14 +1,17 @@
 
 #!/bin/sh
 
-echo "Starting"
+printf "Starting CI presubmit\n"
 
-npm -v
-node -v
-
-BUILDROOT=${BUILDROOT:-github/apigeetool-node}
-
-(cd $BUILDROOT; npm install && npm test)
+(docker build -t apigeetool-test .)
 testStatus=$?
+
+# printf "npm version %s\n" $(npm -v)
+# printf "node version %s\n" $(node -v)
+# 
+# BUILDROOT=${BUILDROOT:-github/apigeetool-node}
+# 
+# (cd $BUILDROOT; npm install && npm test)
+# testStatus=$?
 
 exit ${testStatus}
